+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [next]
 
+### Added
+
+- Add `bundledPlugin(id: String)` helper to IntelliJPlatformPluginsExtension
+- 
 ### Fixed
 
 - Use `withRandomSuffix` for `intellijPlatformConfigurationName` in local dependency configurations defined for Plugin Verifier